requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]==3.12
pymongo
scikit-learn
mlflow
pyaml
dagshub
fastapi
uvicorn
python-multipart

# -e.